+++ src-opam/linux.ml
@@ -31,11 +31,11 @@ let sudo_nopasswd = "ALL=(ALL:ALL) NOPASSWD:ALL"
 
 (** RPM rules *)
 module RPM = struct
-  let update = run "yum update -y"
-  let install fmt = ksprintf (run "yum install -y %s && yum clean all") fmt
+  let update = run "dnf update -y"
+  let install fmt = ksprintf (run "dnf install -y %s && dnf clean all") fmt
 
   let groupinstall fmt =
-    ksprintf (run "yum groupinstall -y %s && yum clean all") fmt
+    ksprintf (run "dnf groupinstall -y %s && dnf clean all") fmt
 
   let add_user ?uid ?gid ?(sudo = false) username =
     let uid = match uid with Some u -> sprintf "-u %d " u | None -> "" in