+++ CHANGES.md
@@ -1,6 +1,7 @@
 unreleased
 ----------
 
+- Switch from yum to DNF for Fedora, CentOS and OracleLinux. (@MisterDA, #???)
 - Add Fedora 36. (@MisterDA #125)
 - Add Ubuntu 22.10. (@MisterDA #124)
 - Support STOPSIGNAL instruction. (@MisterDA #121, review by @avsm)